pyproject.toml
[project]
name = "document-parser"
version = "0.0.1"
description = "Document Parser to extract the tables from PDF documents desinged for - Report Generator Agentic AI"
authors = [
    { name = "Wai Lin Kyaw", email = "wailink13@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.11,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "transformers (>=4.49.0,<5.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "timm (>=1.0.15,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic (>=2.10.6,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "ultralytics (>=8.2.34,<8.3.0)",
    "hypothesis (>=6.130.3,<7.0.0)",
    "pytest (>=8.3.5,<9.0.0)"
]

[tool.poetry]
package-mode=false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
